TST: collect indexing tests by method (#39980)
